requirements.txt
feedparser
requests
beautifulsoup4
schedule
python-dotenv
